Continued Accumulator troubleshooting
Continued messing around with the accumulator setup and the 4 cylinder crankshaft model. I was able to fix the algebraic loop on the relay between the directional valves by using a memory block. Currently the issue with accumulator is that it reads as a pressure below absolute zero at some point, causing the simulation to crash.